+++ CHANGELOG.md
@@ -4,6 +4,8 @@ We follow [Semantic Versions](https://semver.org/).
 
 ## unreleased
 
+- Improve `django.createsuperuser` command with verbose name settings
+
 ## 0.7.1
 
 - Add `recompile-messages` command to django invocations